[chore] automated release of Utility sources
