+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="3.1.0"></a>
+# [3.1.0] - 2025-11-18
+
+## Features
+- Update formatting
+- Update tests
+
+
 <a name="3.0.1"></a>
 # [3.0.1] - 2025-07-11
 
@@ -8,6 +16,7 @@
 - Correct URL in header
 - Fix Unit tests
 
+
 <a name="3.0.0"></a>
 # [3.0.0] - 2025-06-04
 
@@ -15,5 +24,6 @@
 - Release the module
 
 
-[Unreleased]: https://github.com/nice-devone/nice-cxone-mobile-guide-utility-ios/compare/3.0.1...HEAD
+[Unreleased]: https://github.com/nice-devone/nice-cxone-mobile-guide-utility-ios/compare/3.1.0...HEAD
+[3.1.0]: https://github.com/nice-devone/nice-cxone-mobile-guide-utility-ios/compare/3.0.1...3.1.0
 [3.0.1]: https://github.com/nice-devone/nice-cxone-mobile-guide-utility-ios/compare/3.0.0...3.0.1
