+++ Tests/Logging/LogLevelTests.swift
@@ -18,15 +18,13 @@ import Testing
 
 struct LogLevelTests {
 
-    @Test(
-        arguments: zip(LogLevel.allCases, ["TRACE", "DEBUG", "INFO", "WARNING", "ERROR", "FATAL"])
-    )
-    func testDescription(level: LogLevel, label: String) {
+    @Test(arguments: zip(LogLevel.allCases, ["TRACE", "DEBUG", "INFO", "WARNING", "ERROR", "FATAL"]))
+    func verifyDescription(level: LogLevel, label: String) {
         #expect(level.description == label)
     }
 
     @Test(arguments: LogLevel.allCases, LogLevel.allCases)
-    func testComparison(left: LogLevel, right: LogLevel) {
+    func verifyComparison(left: LogLevel, right: LogLevel) {
         #expect(
             (left < right) == (LogLevel.allCases.firstIndex(of: left)! < LogLevel.allCases.firstIndex(of: right)!)
         )
